Renaming package to `model` [skip ci]
