+++ java/server/src/org/openqa/selenium/grid/distributor/local/LocalDistributor.java
@@ -30,7 +30,7 @@ import org.openqa.selenium.grid.data.DistributorStatus;
 import org.openqa.selenium.grid.data.NodeRejectedEvent;
 import org.openqa.selenium.grid.data.NodeStatus;
 import org.openqa.selenium.grid.distributor.Distributor;
-import org.openqa.selenium.grid.distributor.gridmodel.NodeSelector;
+import org.openqa.selenium.grid.distributor.model.NodeSelector;
 import org.openqa.selenium.grid.log.LoggingOptions;
 import org.openqa.selenium.grid.node.Node;
 import org.openqa.selenium.grid.server.BaseServerOptions;
