+++ java/server/src/org/openqa/selenium/grid/distributor/model/BUILD.bazel
@@ -1,7 +1,7 @@
 load("@rules_jvm_external//:defs.bzl", "artifact")
 
 java_library(
-    name = "gridmodel",
+    name = "model",
     srcs = glob(["*.java"]),
     visibility = [
         "//java/server/src/org/openqa/selenium:__subpackages__",
