+++ java/server/test/org/openqa/selenium/grid/distributor/local/BUILD.bazel
@@ -17,7 +17,7 @@ java_test_suite(
         "//java/server/src/org/openqa/selenium/grid/node/local",
         "//java/server/src/org/openqa/selenium/grid/sessionmap/local",
         "//java/server/test/org/openqa/selenium/grid/testing",
-        "//java/server/src/org/openqa/selenium/grid/distributor/gridmodel",
+        "//java/server/src/org/openqa/selenium/grid/distributor/model",
         artifact("com.google.guava:guava"),
         artifact("io.opentelemetry:opentelemetry-api"),
         artifact("junit:junit"),
