+++ java/server/test/org/openqa/selenium/grid/distributor/model/NodeSelectorTest.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.openqa.selenium.grid.distributor.gridmodel;
+package org.openqa.selenium.grid.distributor.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
